requirements.txt
flask
openai 
promptflow
promptflow-azure
autogen
autogen-agentchat~=0.2
azure-core
azure-ai-projects 
azure-ai-inference 
azure-identity 
azure-ai-inference[opentelemetry]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp